server/pyproject.toml
[project]
name = "twilio-realtime-voice"
version = "0.1.0"
description = "Twilio-powered real-time voice agent with Azure Voice Live and web clients"
requires-python = ">=3.9"

dependencies = [
    "quart>=0.20.0",
    "websockets>=15.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
    "quart-cors>=0.8.0",
    "aiohttp>=3.12.14",
    "azure-identity>=1.23.0",
    "azure-storage-blob>=12.22.0",
    "azure-data-tables>=12.7.0",
    "azure-storage-queue>=12.10.0",
    "twilio>=9.3.0",
    "openai[realtime]>=1.93.0",
    "h2>=4.3.0"
]

[tool.uv]
# Define the virtual environment location for uv
project_environment = ".venv"